Added Character tags and Scripts
I added a script to the soldier for movement and varaibles for damage and Health.

I didn't add a script for the  for the tank since I need to implement how the control will shift from character to character.

I am also trying to implement a proper damage and health system