Added Shooting
I managed to get bullets to fly out of the player character. They can only go towards the right and torrent the enemy like desert storm.

Later on I have to figure out how to pull back on how fast the player shoots.

I'm also trying to get a sort of health bar working for both the player and the enemies.